# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TrustPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /disputes/buyer-proof/{id}:
    post:
      operationId: submitBuyerProof
      tags: [disputes]
      summary: Submit buyer proof for a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyerProofBody"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /disputes/seller-proof/{id}:
    post:
      operationId: submitSellerProof
      tags: [disputes]
      summary: Submit seller proof for a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerProofBody"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /disputes/admin/resolve/{id}:
    post:
      operationId: adminResolveDispute
      tags: [admin, disputes]
      summary: Resolve a dispute (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResolveDisputeBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    BuyerProofBody:
      type: object
      properties:
        bankStatementUrl:
          type: string
        txHistoryUrl:
          type: string
    SellerProofBody:
      type: object
      properties:
        bankStatementUrl:
          type: string
        recordingUrl:
          type: string
        lastTxnScreenshotUrl:
          type: string
      required: [bankStatementUrl, recordingUrl, lastTxnScreenshotUrl]
    AdminResolveDisputeBody:
      type: object
      properties:
        winner:
          type: string
          enum: [buyer, seller]
        notes:
          type: string
      required: [winner]