# lib/api-zod/src/generated/types/buyerProofBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BuyerProofBody {
  bankStatementUrl?: string;
  txHistoryUrl?: string;
}